% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{append_database}
\alias{append_database}
\title{Append to the file-list database}
\usage{
append_database(x, path, filename = "database")
}
\arguments{
\item{x}{the tibble or data.frame database}

\item{path}{character the directory to where the database should reside}

\item{filename}{character, the name of the database file}
}
\value{
a tibble with appended data
}
\description{
Append to the file-list database
}
