% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{as_stars}
\alias{as_stars}
\title{Cast a tidync object to stars}
\usage{
as_stars(x, ...)
}
\arguments{
\item{x}{tidync object - typically filtered with \link[tidync:activate]{tidync::activate}
and suitably activated grids with \link[tidync:activate]{tidync::activate}}

\item{...}{other arguments (ignored)}
}
\value{
stars object
}
\description{
Cast a tidync object to stars
}
