% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{attr_fun}
\alias{attr_fun}
\title{A function to apply a function (like \code{log10}) to an attribute of a stars
object.}
\usage{
attr_fun(x, att = "chlor_a", fun = log10, ...)
}
\arguments{
\item{x}{stars object}

\item{att}{character, one attribute name}

\item{fun}{a function that can be applied to the attribute}

\item{...}{other arguments to \code{fun}}
}
\value{
the update inpout oject
}
\description{
A function to apply a function (like \code{log10}) to an attribute of a stars
object.
}
