% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpg.R
\name{browse_obpg}
\alias{browse_obpg}
\title{Browse the metadata OPeNDAP page for a given URL}
\usage{
browse_obpg(x = obpg_url())
}
\arguments{
\item{x}{table with 'meta' and 'opendap' variables.}
}
\value{
the output, invisibly
}
\description{
Browse the metadata OPeNDAP page for a given URL
}
