% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{build_database}
\alias{build_database}
\title{Given a directory, build a database}
\usage{
build_database(path = ".", pattern = glob2rx("*.tif"), save_db = FALSE, ...)
}
\arguments{
\item{path}{a path to raster files}

\item{pattern}{regular expression of the file pattern to search for}

\item{save_db}{logical, if TRUE then write the database}

\item{...}{arguments passed to write_database \strong{if} \code{save_db} is \code{TRUE}.}
}
\value{
tibble or NULL
}
\description{
Given a directory, build a database
}
