% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{compose_obpg}
\alias{compose_obpg}
\alias{compose_filename}
\title{Given a database, build filenames ala \code{yyyy/mmdd/filenames}}
\usage{
compose_obpg(
  x,
  path = ".",
  ext = ".tif",
  from_scratch = FALSE,
  include_nrt = TRUE
)

compose_filename(x, ...)
}
\arguments{
\item{x}{a tibble database}

\item{path}{a path to raster files typically through opbg2/platform/region}

\item{ext}{character - the extension to apply - by default ".tif"}

\item{from_scratch}{logical, if \code{TRUE}, build the name from components}

\item{include_nrt}{logical, if \code{TRUE} then include the NRT flag when building from scratch.
Ignored if from_scratch is \code{FALSE}}

\item{...}{other arguments for \code{compose_obpg}}
}
\value{
character, filenames
}
\description{
Given a database, build filenames ala \code{yyyy/mmdd/filenames}
}
