% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{crop_obpg}
\alias{crop_obpg}
\title{Pseudo-crop a tidync object using a \link[sf:st_bbox]{sf::st_bbox} (or object from which we can determine a \link[sf:st_bbox]{sf::st_bbox})}
\usage{
crop_obpg(x = open_obpg(), y = pnw_bb())
}
\arguments{
\item{x}{tidync object}

\item{y}{object from which a \link[sf:st_bbox]{sf::st_bbox} can be derived}
}
\value{
the output of \link[tidync:hyper_filter]{tidync::hyper_filter}
}
\description{
Note that the input is "hyper-filtered" not actually cropped. This
establishes limits on the grids (dimensions) for subsequent extractions.
One can always restore the original grid dimensions.
}
