% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_missing_days}
\alias{db_missing_days}
\title{Given a database with one or more parameters retrieve a listing of missing dates in the
sequence form first date to last date.}
\usage{
db_missing_days(x = read_database(), last_date = Sys.Date() - 1)
}
\arguments{
\item{x}{a tibble database}

\item{last_date}{either NULL or a Date object to use as the last date of the
date range to examine. If NULL then the last date in the database is used.
By default we use 'yesterday'}
}
\value{
named list of date vectors of mising days (possibly empty)
}
\description{
Given a database with one or more parameters retrieve a listing of missing dates in the
sequence form first date to last date.
}
