% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_most_recent}
\alias{db_most_recent}
\title{Retrieve a subset of most recent entries in a database}
\usage{
db_most_recent(x, n = 1)
}
\arguments{
\item{x}{the database}

\item{n}{numeric, the number of most recent items}
}
\value{
most_recent n of the database by param/per/date
}
\description{
Retrieve a subset of most recent entries in a database
}
