% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_needs_update}
\alias{db_needs_update}
\title{Given per parameter missing days listing, generate a named logical vector
indicating which are missing.}
\usage{
db_needs_update(x = db_missing_days(), last_date = Sys.Date() - 1)
}
\arguments{
\item{x}{a named list of the missing days per parameter generated by
\code{\link{db_missing_days}}}

\item{last_date}{either NULL or a Date object to use as the last date of the
date range to examine. If NULL then the last date in the database is used.
By default we use 'yesterday'.}
}
\value{
a named logical vector indicating which parameters need updating
}
\description{
Given per parameter missing days listing, generate a named logical vector
indicating which are missing.
}
