% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_obpg}
\alias{decompose_obpg}
\alias{decompose_filename}
\title{Decompose one or more obpg filenames into the obpg database}
\usage{
decompose_obpg(
  x = c("AQUA_MODIS.20191130.L3m.DAY.NSST.sst.4km.NRT.nc",
    "NPP_VIIRS.20190703.L3m.DAY.SST.sst.4km.nc",
    "AQUA_MODIS.20030401_20110430.L3m.MC.CHL.chlor_a.9km.nc",
    "AQUA_MODIS.20020701_20020731.L3m.MO.CHL.chlor_a.4km.nc",
    "AQUA_MODIS.20220301_20220331.L3m.MO.CHL.chlor_a.4km.NRT.nc",
    "AQUA_MODIS.20230401_20230430.L3m.MO.POC.poc.4km.NRT.nc")
)

decompose_filename(x, ...)
}
\arguments{
\item{x}{character, vector of filenames to decompose. Path is discarded.}

\item{...}{other arguments for \code{decompose_obpg}}
}
\value{
a database table
\itemize{
\item{date, Date  We only use L3 which is by day, so no POSIXct}
\item{year, numeric date formatted as YYYY}
\item{mmdd, character date formatted as mmdd}
\item{mit, character mission-instrument-type}
\item{level, character string indicating the level}
\item{period, charcater period indicator for L3}
\item{suite, character suite identifier}
\item{param, character parameter name chlor_a, sst, etc}
\item{per, compositing period such as DAY, MO, 8D}
\item{res, character resolution in km}
\item{nrt, character optional Near Real-Time identifier}
\item{file, basename of the input file less the extension}
}
}
\description{
Decompose one or more obpg filenames into the obpg database
}
