% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{explain_database}
\alias{explain_database}
\title{Retrieve an expanatory table about the database}
\usage{
explain_database()
}
\value{
tibble of descriptors (name, type and description) for each field
}
\description{
Retrieve an expanatory table about the database
}
