% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract.default}
\alias{extract.bbox}
\alias{extract.sf}
\alias{extract.sfc_POINT}
\alias{extract.sfc_POLYGON}
\title{extract generic}
\usage{
extract(x, ...)

\method{extract}{default}(x, y = NULL, ...)

\method{extract}{bbox}(x, y = NULL, varname = obpg_vars(y)[1], flip = "y", ...)

\method{extract}{sf}(x, y = NULL, varname = obpg_vars(y), verbose = FALSE, ...)

\method{extract}{sfc_POINT}(x, y = NULL, varname = obpg_vars(y), verbose = FALSE, ...)

\method{extract}{sfc_POLYGON}(x, y = NULL, varname = obpg_vars(y)[1], flip = "y", ...)
}
\arguments{
\item{x}{\code{sfc} object that defines a bounding box}

\item{...}{Arguments passed to or from other methods}

\item{y}{\code{ncdf4} object}

\item{varname}{character one or more variable names}

\item{flip}{char one of "y", "x" or "none" to flip the raster}

\item{verbose}{logical, output helpful messages?}
}
\value{
data frame of covariates (point, or raster)

stars object (one variable per covariate)

tibble of extracted values (one variable per covariate)

stars object (one variable per covariate)
}
\description{
extract generic
}
\section{Methods (by class)}{
\itemize{
\item \code{extract(default)}: Extract data from a NCDF4 object

\item \code{extract(bbox)}: Extract from a NCDF4 object using any sf object

\item \code{extract(sf)}: Extract from a NCDF4 object using any sf object

\item \code{extract(sfc_POINT)}: Extract data from a NCDF4 object using sf POINT object

\item \code{extract(sfc_POLYGON)}: Extract data from a NCDF4 object using sf POLYGON object

}}
