% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_obpg}
\alias{fetch_obpg}
\title{Given a URL, fectch and download to a specified path the data as TIFF.}
\usage{
fetch_obpg(x = obpg_url(), bb = NULL, root = ".")
}
\arguments{
\item{x}{url-database or tidync object with database ("db") attribute}

\item{bb}{NULL or object from which \link[sf:st_bbox]{sf::st_bbox} can be extracted.}

\item{path}{chr, the root data path}
}
\value{
a file database
}
\description{
Subsetting by bounding box is optional
}
