% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_geometry_dimension}
\alias{get_geometry_dimension}
\title{Get geometry dimension code}
\usage{
get_geometry_dimension(x, recursive = FALSE)
}
\arguments{
\item{x}{sf or sfc object}

\item{recursive}{logical, if TRUE drill down to get the type for each
feature.}
}
\value{
character vector such as "XY" or "XYZ"
}
\description{
Get geometry dimension code
}
