% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpg-package.R
\docType{package}
\name{ghrsst}
\alias{ghrsst}
\title{obpg: Simplified access to OBPG (Ocean Biology Processing Group) data from R.}
\description{
Simplified access to OBPG (Ocean Biology Processing Group) data from R.
}
