% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{guess_along}
\alias{guess_along}
\title{Given a database, guess which type of \code{along} might be most appropriate}
\usage{
guess_along(x)
}
\arguments{
\item{x}{database (a tibble)}
}
\value{
either a list of NA_integer_
}
\description{
If the database has only one type of param, then along is a list of times to indicate
laters for one attribute, otherwise along is \code{NA} to indicate multiple attributes.
}
