% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{http_bad}
\alias{http_bad}
\title{Given a URL, test if the server throws an error}
\usage{
http_bad(x)
}
\arguments{
\item{x}{chr, URL to test}
}
\value{
logical, TRUE if an error is thrown
}
\description{
Given a URL, test if the server throws an error
}
