% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_path}
\alias{make_path}
\title{Make a directory as needed}
\usage{
make_path(x, recursive = TRUE, ...)
}
\arguments{
\item{x}{directory name(s)}

\item{recursive}{logical, if TRUE builds missing path segments as needed}

\item{...}{other arguments for \link[base:files2]{base::dir.create}}
}
\value{
named logical vector (TRUE for success(s))
}
\description{
Make a directory as needed
}
