% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpg-package.R
\docType{package}
\name{obpg}
\alias{obpg}
\alias{_PACKAGE}
\title{obpg: Simplified access to OBPG (Ocean Biology Processing Group) data from R.}
\description{
Simplified access to OBPG (Ocean Biology Processing Group) data from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/obpg}
  \item Report bugs at \url{https://github.com/BigelowLab/obpg/issues}
}

}
