% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpg.R
\name{obpg_make_nrt}
\alias{obpg_make_nrt}
\title{Given an opgp_url table, convert to NRT}
\usage{
obpg_make_nrt(x = obpg_url())
}
\arguments{
\item{x}{output of obpg_url()}
}
\value{
table of NRT urls
}
\description{
Given an opgp_url table, convert to NRT
}
