% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{obpg_nc_nav_point}
\alias{obpg_nc_nav_point}
\title{Retrieve obpg navigation values (start, count, lons, lats)}
\usage{
obpg_nc_nav_point(
  X,
  g,
  res = obpg_res(X),
  time = c(1, -1),
  varname = obpg_vars(X)
)
}
\arguments{
\item{X}{ncdf4 object}

\item{g}{geometry object that defines point locations}

\item{res}{numeric, 2 element resolution \code{[res_x,res_y]}}

\item{time}{numeric two elements time indexing \code{[start, length]}.
\code{start} is a 1-based index into the time dimension
\code{length} is the number of indices to retrieve (assumed to be contiguous sequence)}

\item{varname}{character the name of the variable}
}
\value{
data frame 
\itemize{
  \item{g the requested lonlats}
  \item{res the resolution}
  \item{start vector of start values for \code{\link[ncdf4]{ncvar_get}}}
  \item{count vector of count values \code{\link[ncdf4]{ncvar_get}}}
  \item{ext vector of extent see \code{\link[raster]{raster}}}
  \item{crs character, proj string for \code{\link[raster]{raster}}}
  \item{varname character}
}
}
\description{
Retrieve obpg navigation values (start, count, lons, lats)
}
