% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpg.R
\name{obpg_path}
\alias{obpg_path}
\title{Given an object from which a file database can be extracted,
construct a path for writing files.}
\usage{
obpg_path(x, root = ".")
}
\description{
Given an object from which a file database can be extracted,
construct a path for writing files.
}
