% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpg.R
\name{obpg_test_http}
\alias{obpg_test_http}
\title{Test the http service for one or more rows of the output
of \code{\link{obpg_url}}}
\usage{
obpg_test_http(x)
}
\arguments{
\item{x}{table produced by \code{\link{obpg_url}}}
}
\value{
the input table, with possible midifications to the \emph{error}
variable.  TRUE means the http check returns an error code,
FALSE means no error, and NA means untested.
}
\description{
Test the http service for one or more rows of the output
of \code{\link{obpg_url}}
}
