% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{obpg_vars}
\alias{obpg_vars}
\title{Retrieve obpg variables}
\usage{
obpg_vars(X, drop = c("palette"))
}
\arguments{
\item{X}{ncdf4 object}

\item{drop}{character vector of variables to drop}
}
\value{
character vector
}
\description{
Retrieve obpg variables
}
