% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pnw_bb}
\alias{pnw_bb}
\title{Retrieve a bounding box for the Pacific Northwest}
\usage{
pnw_bb(form = c("vector", "bbox", "polygon")[3])
}
\arguments{
\item{form}{chr, one of "vector", "bbox" or "polygon" to determine the
return type}
}
\value{
named vector, sf bbox or sf polygon as requested
}
\description{
Retrieve a bounding box for the Pacific Northwest
}
