% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_obpg}
\alias{read_obpg}
\title{Read raster formatted OBPG file(s).}
\usage{
read_obpg(
  x,
  path = NULL,
  along = guess_along(x),
  tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{obpg database}

\item{path}{character path}

\item{along}{see \link[stars:read_stars]{stars::read_stars} and \link{guess_along}}

\item{tolerance}{num, for binding to allow for some slop}
}
\value{
stars object
}
\description{
Read raster formatted OBPG file(s).
}
