% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{tidync_source}
\alias{tidync_source}
\title{Retrieve the tidync data source}
\usage{
tidync_source(x)
}
\arguments{
\item{x}{tidync object}
}
\value{
filename or url
}
\description{
Retrieve the tidync data source
}
