% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{transform_summary}
\alias{transform_summary}
\title{Retrieve a summary table of each spatial (x and y) transform
for the \strong{selected} space.}
\usage{
transform_summary(x)
}
\arguments{
\item{x}{a \link[tidync:tidync]{tidync::tidync} object}
}
\value{
a table providing a spatial summary
}
\description{
The meaning \strong{selected} space is that defined by \link[tidync:hyper_filter]{tidync::hyper_filter},
defaulting to the full spatial extent if no prior \code{hyper_filter}ing has been done.
}
