% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpg.R
\name{obpg_url}
\alias{obpg_url}
\title{Craft a OBPG URL for a given date}
\usage{
obpg_url(
  date = Sys.Date() - 2,
  where = "opendap",
  root = "http://oceandata.sci.gsfc.nasa.gov/opendap",
  level = c("L3", "L3SMI")[2],
  mission = c("MODIS", "S3A", "SNPP", "ADEOS", "SEASTAR", "PACE")[3],
  instrument = c("AQUA", "TERRA", "SEAWIFS", "VIIRS")[4],
  period = c("DAY", "MONTH")[1],
  product = "SST.sst",
  resolution = "9km"
)
}
\arguments{
\item{date}{character, POSIXt or Date the date to retrieve}

\item{where}{character ignored (for now)}

\item{root}{character, the root URL}

\item{level}{character level component of path}

\item{mission}{character mission component of path}

\item{instrument}{character instrument component of path}

\item{period}{character period component of path}

\item{product}{character, provides version and extend info, leave as default}

\item{resolution}{character resolution component of path}
}
\value{
one or more URLs
}
\description{
Not all OBPG products are supported. If there is one missing that you would
like to add, please contact the package maintainer.
}
\seealso{
\href{https://oceandata.sci.gsfc.nasa.gov/opendap/}{OBPG OPeNDAP catalog}
}
