% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpg.R
\name{obpg_url}
\alias{obpg_url}
\title{Craft a OBPG URL for a given date}
\usage{
obpg_url(
  date = Sys.Date() - 2,
  where = "opendap",
  root = file.path("https://oceandata.sci.gsfc.nasa.gov", "opendap/VIIRS/L3SMI/"),
  product = "L3m_DAY_SNPP_CHL_chl_ocx_4km.nc"
)
}
\arguments{
\item{date}{character, POSIXt or Date the date to retrieve}

\item{where}{character ignored (for now)}

\item{root}{character, the root URL}

\item{product}{character, provides version and extend info, leave as default}
}
\value{
one or more URLs
}
\description{
Craft a OBPG URL for a given date
}
