% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{open_obpg}
\alias{open_obpg}
\title{Open a connection to an OPBG resource}
\usage{
open_obpg(x = obpg_url())
}
\arguments{
\item{x}{chr a single row obpg_url table}
}
\value{
a tidync object or NULL
}
\description{
Open a connection to an OPBG resource
}
